+++ services/index/internal/adapter/controller/event/constant.go
