+++ services/index/internal/adapter/controller/http/ping_handler.go
