+++ services/index/internal/adapter/repository/db/error.go
