+++ services/index/internal/app/url_mappings.go
@@ -1,8 +1,8 @@
 package app
 
 import (
-	"github.com/MurmurationsNetwork/MurmurationsServices/services/index/internal/controller/http"
-	"github.com/MurmurationsNetwork/MurmurationsServices/services/index/internal/repository/db"
+	"github.com/MurmurationsNetwork/MurmurationsServices/services/index/internal/adapter/controller/http"
+	"github.com/MurmurationsNetwork/MurmurationsServices/services/index/internal/adapter/repository/db"
 	"github.com/MurmurationsNetwork/MurmurationsServices/services/index/internal/usecase"
 	"github.com/gin-gonic/gin"
 )
